+++ api-objects/AdminPatient.ts
@@ -1,0 +1,16 @@
+export class TestClass {
+  public myNum: number;
+}
+
+export class GenericAdminPatientResponse {
+  public message: string;
+}
+
+export class ResetPasswordResponse {
+  public password: string;
+}
+
+export class UserListItem {
+  public name: string;
+  public type: string;
+}
